+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/UpdateMasterData/ConsumptionMeteringPoints/ChangeTests.cs
@@ -62,26 +62,6 @@ namespace Energinet.DataHub.MeteringPoints.IntegrationTests.UpdateMasterData.Con
         }
 
         [Fact]
-        public async Task Connection_type_is_changed()
-        {
-            await SendCommandAsync(Scenarios.CreateVEProduction() with
-            {
-                ConnectionType = ConnectionType.Installation.Name,
-            }).ConfigureAwait(false);
-
-            var request = CreateUpdateRequest()
-                with
-                {
-                    ConnectionType = ConnectionType.Direct.Name,
-                };
-
-            await SendCommandAsync(request).ConfigureAwait(false);
-
-            AssertMasterData()
-                .HasConnectionType(ConnectionType.Direct);
-        }
-
-        [Fact]
         public async Task Scheduled_meter_reading_date_is_changed()
         {
             await SendCommandAsync(Scenarios.CreateConsumptionMeteringPointCommand() with
